refactor(type.ts and observer): change observer to TS, update structure type.ts
