+++ src/services/observer.ts
@@ -1,19 +1,27 @@
+import { Component } from 'react';
+
+import { DataForCreateCharts } from '@/pages/Timeline/ChartCurrency';
+
 class Observer {
+  data: DataForCreateCharts | null;
+
+  observers: Component[];
+
   constructor() {
     this.data = null;
     this.observers = [];
   }
 
-  setData(data) {
+  setData(data:DataForCreateCharts) {
     this.data = data;
     this.notify();
   }
 
-  subscribe(observer) {
+  subscribe(observer:Component) {
     this.observers.push(observer);
   }
 
-  unsubscribe(observer) {
+  unsubscribe(observer:Component) {
     this.observers = this.observers.filter((obs) => obs !== observer);
   }
